+++ CHANGELOG.md
@@ -1,15 +1,19 @@
 # Changelogs
 ## Version-1.0 (28-Apr-2020)
-##Feature
+### Feature
 - Option to Detect the algorithm of the input JWT
 - Base64 decode the input JWT Token
 - Generate new JWT Token by changing algorthim to `None` and `HS256/384/512`
 
 ## Version-1.1 (29-Apr-2020)
-##Feature
+### Feature
 - Exception Handling implemented
 
 ## Version-1.1.1 (30-Apr-2020)
-##Feature
+### Feature
 - Bug fixes
 
+## Version-1.5 (24-May-2020)
+### Feature
+- Generate Public and Private key pair, if not provided as input for signing
+- Generate new JWT Token by changing algorthim to `RS256/384/512`